Catch error while saving game data
If there is some issue saving the game data, this should be logged but
it should not crash the game.
This is important as an exception in that function seems to result in a
bad state of the server (the game is not cleaned up properly and players
are not added back to the queue).