Matchmaking: Sample from all possible matches
Before this commit, the matchmaking was looping through the queue and
matching a player with the first candidate where the match quality was
above the threshold.
This commit change this so that match qualities are computed for all
other candidates in the queue and then a match is found by sampling form
all options above the threshold.

Use caching for the match quality score, so it is only computed once for
each pair of user.  This is probably not really a concern
performance-wise but has the benefit that only one entry is added to the
monitoring data.