Show client configuration in the user dashboard
To make it easier for users to figure out how to configure their client,
add a code block setting the environment variables to the corresponding
values (so they can simply copy-paste that block).
Also add a link to the GitHub repo with the example client code.

Add a new config value "server_url" for this, so we don't need to
hard-code the server.