Periodically write monitoring information to a file
The information included in the file are
- currently connected players
- currently running games
- the matchmaking queue
- matchmaking scores

With this, it is possible to quickly check what is currently going on
without the need to add some extra interface to the server (which would
likely the nicer but more complicated approach).  I recommend using a
file in `/dev/shm` for this to avoid unnecessary IO operations.