Add limit for number of parallel games
The limit can be configured via the setting `max_parallel_games`.
Matchmaking simply returns early if the limit is reached.  Note that
this results in missing or incomplete match quality scores when the
limit is reached but I prefer this over a more complicated
implementation (those scores are anyway only relevant when games can be
started).