Add script to render games based on recorded observations
Load observations from a recorded game and set them to the environment,
rendering every step.

If a database is provided as argument, it is used to get info about
player names and final score, which is added in the images.

The result can either be viewed directly or saved as a video.