Randomise order of players in a game
Depending on number of players and matchmaking scores, some players may
end up being the first player more often (the longer you wait for a
match, the higher up you are in the queue and thus the more likely you
are to be the first player).

To avoid issues with games where the winning changes are not equal for
first and second player, randomize the order of players when starting
the game.