Remove option to swap players in video playback
There is no swapping in the game anymore, so this option is not needed
anymore.  The script is anyway not compatible with old game files
anymore, so no need to keep the option for backward compatibility.

If swapping is ever to be introduced in the game again, the information
should be added to the game file instead of relying on the user to pass
it as argument when playing back the file.