Update render_game.py to new game data format
The data stored for each game changed in recent commits.  This commit
adapts the render_game.py script to work with the new files.

Since all relevant information is now stored directly in the game file,
no lookup in the database is needed anymore.  Hence, the option to
specify a database is removed.

BREAKING: This is a breaking change, the script won't work with old game
files anymore!