Use `mu - 3*sigma` for ranking score
This is what Openskill (and I think also Trueskill) are using, so let's
adopt it.
It has the nice side effect that new users start with a score of zero
instead of some random number.

I also refactored a bit to not have that expression repeated everywhere.
It is now only defined in two functions in the `User` class (one for the
order-by expression, one for computing the actual score).  This will
make it easier to change the expression if the future if needed.

I hope I got all places where `mu - sigma` was used.  At least by
greping the source for "sigma", I didn't find any other occurrences.