Always read matchmaking parameters directly from config
For config reloading to make sense, parameters that can be reloaded
should always be read directly from the config structure (so that
changes there actually take effect).

This commit changes the MatchmakingManager accordingly, so that the
corresponding config values are no longer copied to class variables but
instead it always reads them from the up-to-date config (actually, the
config is only updated once per function call, so changes during a
function execution only take effect in the next iteration).