Adapt Dockerfile to more easily send the "reload" signal
Normally Docker will run the CMD in a shell like this: `sh -c
'comprl-server ...'`.  This has the problem that when sending a signal
to the container (e.g. with `docker compose kill ...`), the signal ends
up at the `sh` process, not the `comprl-server` process.
By using `exec`, `comprl-server` will become the main process and thus
receive the signals directly.

For convenience, also add a "reload" target to the makefile that sends
the SIGHUP signal for reloading configuration to the container.