Enforce unique access tokens in the database.
We are using UUIDs for the tokens, which in practice should never
collide, but since it's trivial to enforce by simply making the db field
unique, let's do that.
In the (very very VERY) unlikely case of two users with the same token,
user creation should now fail with some error.  The error is no handled
anywhere, so it might crash something but that's still better than the
confusion that would arise from two players with the same token.